src/main/java/com/hhy/sell/service/impl/CategoryServiceImpl.java
package com.hhy.sell.service.impl;

import com.hhy.sell.dataobject.ProductCategory;
import com.hhy.sell.repository.ProductCategoryRepository;
import com.hhy.sell.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @author OceansHan
 * @date 2020/9/10 14:54
 */
@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {

        return repository.findById(categoryId).orElse(null);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryByTypeList) {
        return repository.findByCategoryTypeIn(categoryByTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
